% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglemanise_utils.R
\name{angl_cor}
\alias{angl_cor}
\title{Adjusted big_cor function with different scaling function.
In comparison to big_scale(), we don't stop the scaling if there is a gene
with 0 variance.
#' Correlation}
\usage{
angl_cor(X, block.size = 1000)
}
\arguments{
\item{X}{An object of class \link[=FBM-class]{FBM}.}

\item{block.size}{Number of columns to process at a time.}
}
\value{
An FBM object with the (pearson) correlation matrix of the data in X.
}
\description{
Compute the correlation matrix of a Filebacked Big Matrix.
}
\examples{
X <- bigstatsr::FBM(13, 17, init = rnorm(221))
angl_cor(X)
}
