% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{anglem-methods}
\alias{anglem-methods}
\alias{show,anglem-method}
\alias{matrix_list}
\alias{matrix_list<-}
\alias{dataset_key}
\alias{batch_key}
\alias{data_info}
\alias{angl_weights}
\alias{angl_weights<-}
\alias{list_stats}
\alias{list_stats<-}
\alias{intersect_genes}
\alias{intersect_genes<-}
\alias{extract_integration_genes}
\alias{add_unique_batch_key}
\title{Anglem Class for Storing and Processing Gene Expression Data}
\usage{
\S4method{show}{anglem}(object)

matrix_list(object)

matrix_list(object) <- value

dataset_key(object)

batch_key(object)

data_info(object)

angl_weights(object)

angl_weights(object) <- value

list_stats(object)

list_stats(object) <- value

intersect_genes(object)

intersect_genes(object) <- value

extract_integration_genes(object)

add_unique_batch_key(
  seurat_object,
  dataset_key = NA_character_,
  batch_key,
  new_unique_batch_key = "batch"
)
}
\arguments{
\item{object}{An \code{anglem} object.}

\item{value}{A character vector of gene names.}

\item{seurat_object}{A \code{\link[Seurat]{Seurat}} object.}

\item{dataset_key}{A character string specifying the column name in the
metadata that identifies the dataset. If \code{NA}, only the
\code{batch_key} is used.}

\item{batch_key}{A character string specifying the column name in the
metadata that identifies the batch.}

\item{new_unique_batch_key}{A character string for the new unique batch key
to be added to the metadata. Default is \code{"batch"}.}
}
\value{
Prints a summary to the console.

A list of \code{\link[bigstatsr]{FBM}} objects containing gene
expression matrices.

The updated \code{anglem} object.

A character string representing the dataset key.

A character string representing the batch key.

A data frame containing dataset and batch information.

A named numeric vector of weights.

The updated \code{anglem} object.

A list containing statistical matrices such as mean z-scores and SNR
z-scores

The updated \code{anglem} object.

A character vector of intersected gene names.

The updated \code{anglem} object.

A character vector of integration gene names.

A \code{\link[Seurat]{Seurat}} object with an additional metadata
column containing the unique batch key.
}
\description{
The \code{anglem} class is designed to construct the correct input for the
\code{anglemanise} function from a \code{\link[Seurat]{Seurat}} object and store
the results of the analysis. It encapsulates the data and metadata required
for processing gene expression data across multiple datasets and batches.

This method provides a concise summary of an \code{anglem} object, including
dataset and batch information, the number of intersected genes, and other
relevant details.

Retrieves the list of gene expression matrices stored in the \code{anglem}
object.

Assigns a new list of gene expression matrices to the \code{anglem} object.

Retrieves the dataset key used in the \code{anglem} object.

Retrieves the batch key used in the \code{anglem} object.

Retrieves the data frame summarizing the selected anglemania gene pairs
\code{anglem} object.

Retrieves the weights assigned to each dataset or batch in the \code{anglem}
object.

Assigns new weights to the datasets or batches in the \code{anglem} object.

Retrieves the list of statistical measures computed across datasets in the
\code{anglem} object.

Assigns a new list of statistical measures to the \code{anglem} object.

Retrieves the vector of genes that are expressed in at least the specified
number of cells across all batches.

Assigns a new vector of intersected genes to the \code{anglem} object.

Retrieves the list of genes selected for integration from the \code{anglem}
object.

This function adds a unique batch identifier to the metadata of a
\code{\link[Seurat]{Seurat}} object by combining specified dataset and batch
keys. This is useful for distinguishing samples during integration or
analysis.
}
\section{Functions}{
\itemize{
\item \code{show(anglem)}: show anglem object info

\item \code{matrix_list()}: Access matrix list

\item \code{matrix_list(object) <- value}: set matrix list in anglem object

\item \code{dataset_key()}: Access dataset key of anglem object

\item \code{batch_key()}: Access batch key of anglem object

\item \code{data_info()}: Access info of selected gene pairs

\item \code{angl_weights()}: Access weights

\item \code{angl_weights(object) <- value}: Set weights

\item \code{list_stats()}: Access statistics of the gene-gene matrices

\item \code{list_stats(object) <- value}: Set statistics of the gene-gene matrices

\item \code{intersect_genes()}: Access the intersection of genes of all batches

\item \code{intersect_genes(object) <- value}: Set the intersection of genes of all batches

\item \code{extract_integration_genes()}: Access the genes extracted by anglemania

\item \code{add_unique_batch_key()}: Temporarily add a unique batch key to the dataset

}}
\section{Slots}{

\describe{
\item{\code{matrix_list}}{A list of \code{\link[bigstatsr]{FBM}} objects containing
the gene expression matrices for each batch.}

\item{\code{dataset_key}}{A character string indicating the key used to denote the
dataset in the metadata.}

\item{\code{batch_key}}{A character string indicating the key used to denote the
batch in the metadata.}

\item{\code{data_info}}{A data frame summarizing the number of samples per dataset
and their weights.}

\item{\code{weights}}{A numeric vector of weights for each dataset based on the
number of samples.}

\item{\code{list_stats}}{A list containing statistical measures computed across the
datasets.}

\item{\code{intersect_genes}}{A character vector of genes that are expressed in at
least the specified number of cells across all batches.}

\item{\code{min_cells_per_gene}}{A numeric value indicating the minimum number of
cells in which a gene must be expressed to be included in the analysis.}

\item{\code{integration_genes}}{A list containing information about integration genes
and their statistics.}
}}

\seealso{
\code{\link{create_anglem}}, \code{\link{big_anglemanise}}
}
