% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglemania-package.R
\docType{package}
\name{anglemania-package}
\alias{anglemania}
\alias{anglemania-package}
\title{anglemania: What the Package Does (One Line, Title Case)}
\description{
anglemania extracts genes from multi-batch scRNA-seq experiments for downstream dataset integration. It improves conventional usage of highly-variable genes for integration tasks.
}
\section{anglemania functions}{

The anglemania functions ...
}

\author{
\strong{Maintainer}: Aaron Kollotzek \email{aaron.kollotzek@mdc-berlin.de} (\href{https://orcid.org/0009-0009-7142-4015}{ORCID})

}
\keyword{internal}
