% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{show,anglemaniaObject-method}
\alias{show,anglemaniaObject-method}
\alias{matrix_list}
\alias{matrix_list,anglemaniaObject-method}
\alias{anglemaniaObject-methods,matrix_list}
\alias{matrix_list<-}
\alias{matrix_list<-,anglemaniaObject-method}
\alias{anglemaniaObject-methods,matrix_list<-}
\alias{dataset_key}
\alias{dataset_key,anglemaniaObject-method}
\alias{anglemaniaObject-methods,dataset_key}
\alias{batch_key}
\alias{batch_key,anglemaniaObject-method}
\alias{anglemaniaObject-methods,batch_key}
\alias{data_info}
\alias{data_info,anglemaniaObject-method}
\alias{anglemaniaObject-methods,data_info}
\alias{angl_weights}
\alias{angl_weights,anglemaniaObject-method}
\alias{anglemaniaObject-methods,angl_weights}
\alias{angl_weights<-}
\alias{angl_weights<-,anglemaniaObject-method}
\alias{anglemaniaObject-methods,angl_weights<-}
\alias{list_stats}
\alias{list_stats,anglemaniaObject-method}
\alias{anglemaniaObject-methods,list_stats}
\alias{list_stats<-}
\alias{list_stats<-,anglemaniaObject-method}
\alias{anglemaniaObject-methods,list_stats<-}
\alias{intersect_genes}
\alias{intersect_genes,anglemaniaObject-method}
\alias{anglemaniaObject-methods,intersect_genes}
\alias{intersect_genes<-}
\alias{intersect_genes<-,anglemaniaObject-method}
\alias{anglemaniaObject-methods,intersect_genes<-}
\alias{get_anglemania_genes}
\alias{get_anglemania_genes,anglemaniaObject-method}
\alias{anglemaniaObject-methods,get_anglemania_genes}
\alias{add_unique_batch_key}
\title{Display Summary Information for an anglemaniaObject}
\usage{
\S4method{show}{anglemaniaObject}(object)

matrix_list(object)

\S4method{matrix_list}{anglemaniaObject}(object)

matrix_list(object) <- value

\S4method{matrix_list}{anglemaniaObject}(object) <- value

dataset_key(object)

\S4method{dataset_key}{anglemaniaObject}(object)

batch_key(object)

\S4method{batch_key}{anglemaniaObject}(object)

data_info(object)

\S4method{data_info}{anglemaniaObject}(object)

angl_weights(object)

\S4method{angl_weights}{anglemaniaObject}(object)

angl_weights(object) <- value

\S4method{angl_weights}{anglemaniaObject}(object) <- value

list_stats(object)

\S4method{list_stats}{anglemaniaObject}(object)

list_stats(object) <- value

\S4method{list_stats}{anglemaniaObject}(object) <- value

intersect_genes(object)

\S4method{intersect_genes}{anglemaniaObject}(object)

intersect_genes(object) <- value

\S4method{intersect_genes}{anglemaniaObject}(object) <- value

get_anglemania_genes(object)

\S4method{get_anglemania_genes}{anglemaniaObject}(object)

add_unique_batch_key(
  seurat_object,
  dataset_key = NA_character_,
  batch_key,
  new_unique_batch_key = "batch"
)
}
\arguments{
\item{object}{An \code{anglemaniaObject}.}

\item{value}{A character vector of gene names.}

\item{seurat_object}{A \code{\link[Seurat]{Seurat}} object.}

\item{dataset_key}{A character string specifying the column name in the
metadata that identifies the dataset. If \code{NA}, only the
\code{batch_key} is used.}

\item{batch_key}{A character string specifying the column name in the
metadata that identifies the batch.}

\item{new_unique_batch_key}{A character string for the new unique batch key
to be added to the metadata. Default is \code{"batch"}.}
}
\value{
Prints a summary to the console.

A list of \code{\link[bigstatsr]{FBM}} objects containing gene
expression matrices.

The updated \code{anglemaniaObject}.

A character string representing the dataset key.

A character string representing the batch key.

A data frame containing dataset and batch information.

A named numeric vector of weights.

The updated \code{anglemaniaObject}.

A list containing statistical matrices such as mean z-scores and SNR
z-scores

The updated \code{anglemaniaObject}.

A character vector of intersected gene
names from multiple Seurat objects.

The updated \code{anglemaniaObject} object.

A character vector of integration gene names.

A \code{\link[Seurat]{Seurat}} object with an additional metadata
column containing the unique batch key.
}
\description{
This method provides a concise summary of an \code{anglemaniaObject}, including
dataset and batch information, the number of intersected genes, and other
relevant details.

Retrieves the list of gene expression matrices stored in the \code{anglemaniaObject}
object.

Assigns a new list of gene expression matrices to the \code{anglemaniaObject}.

Retrieves the dataset key used in the \code{anglemaniaObject}.

Retrieves the batch key used in the \code{anglemaniaObject}.

Retrieves the data frame summarizing the selected anglemania gene pairs
\code{anglemaniaObject}.

Retrieves the weights assigned to each dataset or batch in the \code{anglemaniaObject}

Assigns new weights to the datasets or batches in the \code{anglemaniaObject}.

Retrieves the list of statistical measures computed across datasets in the
\code{anglemaniaObject}.

Assigns a new list of statistical measures to the \code{anglemaniaObject}.

Retrieves the vector of genes that are expressed in at least the specified
number of cells across all batches.

Assigns a new vector of intersected genes to the \code{anglemaniaObject}.

Retrieves the list of genes selected for integration from the \code{anglemaniaObject}

This function adds a unique batch identifier to the metadata of a
\code{\link[Seurat]{Seurat}} object by combining specified dataset and batch
keys. This is useful for distinguishing samples during integration or
analysis.
}
\section{Functions}{
\itemize{
\item \code{show(anglemaniaObject)}: show anglemaniaObject info

\item \code{matrix_list()}: Access matrix list

\item \code{matrix_list(object) <- value}: set matrix list in anglemaniaObject

\item \code{dataset_key()}: Access dataset key of anglemaniaObject

\item \code{batch_key()}: Access batch key of anglemaniaObject

\item \code{data_info()}: Access info of selected gene pairs

\item \code{angl_weights()}: Access weights

\item \code{angl_weights(object) <- value}: Set weights

\item \code{list_stats()}: Access statistics of the gene-gene matrices

\item \code{list_stats(object) <- value}: Set statistics of the gene-gene matrices

\item \code{intersect_genes()}: Access the intersection of genes of all batches

\item \code{intersect_genes(object) <- value}: Set the intersection of genes of all batches

\item \code{get_anglemania_genes()}: Access the genes extracted by anglemania

\item \code{add_unique_batch_key()}: Temporarily add a unique batch key to the dataset

}}
\examples{
se <- SeuratObject::pbmc_small
anglemania_object <- create_anglemaniaObject(se, batch_key = "groups")
anglemania_object <- anglemania(anglemania_object)
show(anglemania_object)
se <- SeuratObject::pbmc_small
anglemania_object <- create_anglemaniaObject(se, batch_key = "groups")
anglemania_object <- anglemania(anglemania_object)
str(matrix_list(anglemania_object))
se <- SeuratObject::pbmc_small
se[[]]$Dataset <- rep(c("A", "B"), each = ncol(se) / 2)
anglemania_object <- create_anglemaniaObject(
  se,
  dataset_key = "Dataset",
  batch_key = "groups",
  min_cells_per_gene = 1
)
dataset_key(anglemania_object)
se <- SeuratObject::pbmc_small
anglemania_object <- create_anglemaniaObject(
  se,
  batch_key = "groups",
  min_cells_per_gene = 1
)
batch_key(anglemania_object)
se <- SeuratObject::pbmc_small
anglemania_object <- create_anglemaniaObject(
  se,
  batch_key = "groups",
  min_cells_per_gene = 1
)
batch_key(anglemania_object)
se <- SeuratObject::pbmc_small
anglemania_object <- create_anglemaniaObject(
  se,
  batch_key = "groups",
  min_cells_per_gene = 1
)
batch_key(anglemania_object)
angl_weights(anglemania_object)
se <- SeuratObject::pbmc_small
anglemania_object <- create_anglemaniaObject(
  se,
  batch_key = "groups",
  min_cells_per_gene = 1
)
anglemania_object <- anglemania(anglemania_object)
stats <- list_stats(anglemania_object)
str(stats)

se <- SeuratObject::pbmc_small
anglemania_object <- create_anglemaniaObject(
  se,
  batch_key = "groups",
  min_cells_per_gene = 1
)
intersect_genes(anglemania_object)
se <- SeuratObject::pbmc_small
anglemania_object <- create_anglemaniaObject(
  se,
  batch_key = "groups",
  min_cells_per_gene = 1
)
anglemania_object <- anglemania(anglemania_object)
# extract the genes identified by anglemania()
anglemania_genes <- get_anglemania_genes(anglemania_object)

se <- SeuratObject::pbmc_small
se[[]]$Dataset <- rep(c("A", "B"), each = ncol(se)/2)
se <- add_unique_batch_key(
  seurat_object = se,
  dataset_key = "Dataset",
  batch_key = "groups",
  new_unique_batch_key = "batch" 
  )
head(se[[]])
}
\seealso{
\code{\link{anglemania}} \code{\link{get_list_stats}}
}
