% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglemanise.R
\name{anglemanise}
\alias{anglemanise}
\title{Record critical angles between genes in gene expression
matrices from a Seurat list.}
\usage{
anglemanise(
  seurat_list,
  extrema = 0.001,
  n_threads = 16,
  path_to_write_angles = "."
)
}
\arguments{
\item{seurat_list}{seurat list.}

\item{extrema}{double. Fraction of the angles
to be cut from both sides of an approximated angle
distribution.}

\item{n_threads}{integer. Number of cores to use for
the computation of the cosine distances.}

\item{path_to_write_angles}{string. A path, under which
a temporary directory will be created to record values
of critical angles for each sample.}
}
\value{
list. First two elements are sparse matrices
recording the number of sharp and blunt angles across
the datasets. Third and fourth elements are lists with
angles statistics and paths to the values of critical
angles.
}
\description{
\emph{anglemanise} is a high-level function that returns a list
of lists containing critical angles between genes across
input samples, angle statistics per sample, paths to the
the angle records per sample.
}
\details{
This is the main function of the package. It calculates
angles between all genes across all samples provided in
the seurat list. It approximates the angle distribution
for each sample and extracts values of critical angles.
}
\seealso{
https://arxiv.org/abs/1306.0256
}
