% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglemanise_utils.R
\name{anglemanise_utils}
\alias{anglemanise_utils}
\alias{sparse_to_fbm}
\alias{get_dstat}
\alias{big_mat_list_mean}
\alias{get_list_stats}
\alias{extract_rows_for_unique_genes}
\alias{prefilter_angl}
\alias{select_genes}
\alias{replace_with_na}
\alias{variable_genes_overlap}
\title{Utility Functions for the anglemania Package}
\usage{
sparse_to_fbm(s_mat)

get_dstat(corr_matrix)

big_mat_list_mean(angl)

get_list_stats(angl)

extract_rows_for_unique_genes(dt, max_n_genes)

prefilter_angl(angl, zscore_mean_threshold = 1, zscore_sn_threshold = 1)

select_genes(
  angl,
  zscore_mean_threshold = 2,
  zscore_sn_threshold = 2,
  max_n_genes = NULL,
  direction = "both",
  adjust_thresholds = TRUE
)

replace_with_na(v)

variable_genes_overlap(
  seu,
  ang_obj,
  zscore_mean_thresholds = c(0.5, 1:15),
  zscore_sn_thresholds = c(0.5, 1:15),
  adjust_thresholds = FALSE,
  layer = "data"
)
}
\arguments{
\item{s_mat}{A sparse matrix.}

\item{corr_matrix}{An \code{\link[bigstatsr]{FBM}} object.}

\item{angl}{An \code{anglemania_object} containing statistical
matrices such as mean z-scores and SNR z-scores.}

\item{dt}{A data frame containing gene pairs, with columns \code{geneA}
and \code{geneB}.}

\item{max_n_genes}{Integer specifying the maximum number of genes to select.
If \code{NULL}, all genes that pass the thresholds are used. Default is
\code{NULL}.}

\item{zscore_mean_threshold}{Numeric value specifying the threshold for the
absolute mean z-score. Default is 2.}

\item{zscore_sn_threshold}{Numeric value specifying the threshold for the
SNR z-score. Default is 2.}

\item{direction}{whether to select genes with positive, negative or both
mean z-scores. Default is "both"}

\item{adjust_thresholds}{Logical. Whether to use adaptive thresholding in
\code{select_genes}. Default is FALSE.}

\item{v}{A numeric vector.}

\item{seu}{A Seurat object. If variable features have not been identified,
they will be computed using \code{\link[Seurat]{FindVariableFeatures}}.}

\item{ang_obj}{An anglemania analysis object. This is used by \code{select_genes}
and \code{get_anglemania_genes} to select genes based on the provided thresholds.}

\item{zscore_mean_thresholds}{A numeric vector of zscore mean thresholds used
for gene selection from \code{ang_obj}. Default is \code{c(0.5, 1:15)}.}

\item{zscore_sn_thresholds}{A numeric vector of zscore signal to noise ratio
thresholds used for gene selection from \code{ang_obj}. Default is
\code{c(0.5, 1:15)}.}

\item{layer}{Character. The assay layer in the Seurat object to use when
calculating variable features (e.g., "counts", "data", or "logcounts").
Default is \code{"data"}.}
}
\value{
An \code{\link[bigstatsr]{FBM}} object from the \pkg{bigstatsr}
package.

A list with statistical measures including \code{mean}, \code{sd},
\code{var}, \code{min}, and \code{max}.

A new \code{\link[bigstatsr]{FBM}} object containing the mean values.

A list containing three matrices: \code{mean_zscore},
\code{sds_zscore}, and \code{sn_zscore}.

A vector of unique gene identifiers.

The input \code{anglemania_object} with the
\code{integration_genes} slot updated to include the selected genes and
their statistical information.

The input \code{anglemaniaObject} with the
\code{integration_genes} slot updated to include the selected genes and
their statistical information.

A numeric vector with NaN and Inf values replaced with NA.

A data frame with the following columns:
\describe{
\item{threshold}{The threshold value used for gene selection.}
\item{intersecting_genes}{The number of genes that are both variable in
\code{seu} and selected by \code{ang_obj}.}
\item{perc_var_genes}{The proportion of variable genes in \code{seu} that
are also selected by \code{ang_obj}.}
\item{number_angl_genes}{The total number of genes selected by
\code{ang_obj} at the given threshold.}
}
}
\description{
A collection of utility functions used within the
\pkg{anglemania} package
for manipulating FBMs, calculating statistics, and selecting genes.

Replaces all NaN and Inf values in a numeric vector with NA.
}
\details{
The function combines the \code{geneA} and \code{geneB} columns, extracts
unique gene names, and returns the first \code{max_n_genes} genes. If
\code{max_n_genes} exceeds the number of unique genes available, all unique
genes are returned.

The function performs the following steps:
\enumerate{
\item Checks if the input object is of class \code{\link{anglemania_object-class}}.
\item Identifies gene pairs where both the mean z-score and SNR z-score
exceed the specified thresholds.
}

The function performs the following steps:
\enumerate{
\item Checks if the input object is of class \code{\link{anglemania_object-class}}.
\item If \code{max_n_genes} is not specified, it uses all genes that pass
the thresholds.
\item Identifies gene pairs where both the mean z-score and SNR z-score
exceed the specified thresholds.
\item If no gene pairs meet the criteria, it adjusts the thresholds to the
99th percentile values of the corresponding statistics and re-selects.
\item Extracts unique genes from the selected gene pairs using
\code{\link{extract_rows_for_unique_genes}}.
\item Updates the \code{integration_genes} slot of the
\code{anglemania_object}
with the selected genes and their statistics.
}
}
\section{Functions}{
\itemize{
\item \code{sparse_to_fbm()}: Convert a sparse matrix into a file-backed matrix
(\code{\link[bigstatsr]{FBM}}) with efficient memory usage.

\item \code{get_dstat()}: Compute mean, standard deviation, variance,
min, and max of a correlation matrix stored as an
\code{\link[bigstatsr]{FBM}}.

\item \code{big_mat_list_mean()}: Calculates the element-wise mean from a list
of \code{\link[bigstatsr]{FBM}}s stored in an \code{anglemania_object}.

\item \code{get_list_stats()}: Calculate mean, standard deviation, and SNR
across a list of FBMs stored in an \code{anglemania_object}.

\item \code{extract_rows_for_unique_genes()}: Extract unique gene identifiers from gene pairs,
returning up to a specified maximum number.

\item \code{prefilter_angl()}: Preselect genes from an
\code{anglemania_object}
based on z-score thresholds, to simplify downstream filtering.

\item \code{select_genes()}: Select genes from an \code{anglemania_object}
based on z-score thresholds for mean and signal-to-noise ratio (SNR).
Selects genes from an \code{\link{anglemania_object-class}} based on specified
thresholds for the absolute mean z-score and signal-to-noise ratio
(SNR) z-score. It updates the \code{integration_genes} slot of the
\code{anglemania_object} with the selected genes and associated
information.

\item \code{replace_with_na()}: replace Nan and Inf values with NA

\item \code{variable_genes_overlap()}: Overlap of Variable Genes Across Thresholds.
This function computes the overlap between variable genes identified in a
Seurat object and the gene sets selected from an "anglemania" analysis
object (\code{ang_obj}) at multiple threshold levels. For each threshold, genes
are selected from \code{ang_obj} by applying \code{select_genes}, and the intersection
of these genes with the variable features of the Seurat object is calculated.
The result is a summary data frame that includes the number and percentage of
variable genes overlapping with the anglemania-selected genes, as well as the
number of genes selected at each threshold.

}}
\examples{
s_mat <- Matrix::rsparsematrix(nrow = 10, ncol = 5, density = 0.3)
fbm_mat <- sparse_to_fbm(s_mat)
fbm_mat
s_mat <- Matrix::rsparsematrix(nrow = 10, ncol = 5, density = 0.3)
fbm_mat <- sparse_to_fbm(s_mat)
result <- get_dstat(fbm_mat)
str(result)
result
# Create FBMs
mat1 <- matrix(1:9, nrow = 3)
mat2 <- matrix(1:9, nrow = 3)

fbm1 <- bigstatsr::FBM(nrow = nrow(mat1), ncol = ncol(mat1), init = mat1)
fbm2 <- bigstatsr::FBM(nrow = nrow(mat2), ncol = ncol(mat2), init = mat2)

# Create weights
weights <- c(batch1 = 0.5, batch2 = 0.5)

# Create the list of FBMs
fbm_list <- list(batch1 = fbm1, batch2 = fbm2)

# Construct the anglemania_object
angl <- new(
  "anglemania_object",
  weights = weights,
  matrix_list = fbm_list
)
big_mat_list_mean(angl)
sce <- sce_example()
angl <- create_anglemania_object(sce, batch_key = "batch")
angl <- anglemania(angl)
list_stats(angl) <- get_list_stats(angl)
str(list_stats(angl))
gene_pairs <- data.frame(
  geneA = c("Gene1", "Gene2", "Gene3", "Gene4"),
  geneB = c("Gene3", "Gene4", "Gene5", "Gene6")
)
unique_genes <- extract_rows_for_unique_genes(
  gene_pairs,
  max_n_genes = 3
)
print(unique_genes)
sce <- sce_example()
angl <- create_anglemania_object(sce, batch_key = "batch")
angl <- anglemania(angl)
angl <- select_genes(angl,
                      zscore_mean_threshold = 2.5,
                     zscore_sn_threshold = 2.5,
                     max_n_genes = 2000)
anglemania_genes <- get_anglemania_genes(angl)
# View the selected genes and use for integration
v <- c(1, 2, 3, 4, 5, 6, 7, Inf, 9, NA)
v <- replace_with_na(v)
v
sce <- sce_example()
seu <- Seurat::CreateSeuratObject(counts = SingleCellExperiment::counts(sce))
seu <- Seurat::FindVariableFeatures(seu)
angl <- create_anglemania_object(
  sce,
  batch_key = "batch",
  min_cells_per_gene = 1
)
angl <- anglemania(angl)
head(variable_genes_overlap(seu, angl))
}
\seealso{
\code{\link[bigstatsr]{big_apply}}, \code{\link[bigstatsr]{FBM}}

\code{\link[bigstatsr]{big_apply}}, \code{\link[bigstatsr]{FBM}}

\code{\link{select_genes}}

\code{\link{extract_rows_for_unique_genes}},
\code{\link{intersect_genes}}, \code{\link{list_stats}}
}
\keyword{internal}
