% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_angles.R
\name{approximate_angles}
\alias{approximate_angles}
\title{Approximate angle distribution}
\usage{
approximate_angles(x_df_ang, quantile_split)
}
\arguments{
\item{x_df_ang}{data.table. A long data.table with three columns:
x - name of a gene in row, y - name of a gene in column,
angle - an angle between x and y.}

\item{quantile_split}{double. minimal splitting unit to
construct a vector of probabilities from 0 to 1.}
}
\value{
list. List with multiple entries, one of which contains
an approxiamted distribution. Others contains NAs to be filled by
subsequent functions.
}
\description{
Split the angle distribution by a vector of probabilities.
}
\details{
Vector of probabilities for a quantile function is constructed
from a minimal splitting unit provided by the user.
}
