% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_conmat.R
\name{assemble_conmat}
\alias{assemble_conmat}
\title{Assemble conserved matrix}
\usage{
assemble_conmat(ohm)
}
\arguments{
\item{ohm}{data.frame. An output from the \strong{assemble_onehotmat} function.}
}
\value{
data.frame. Rows are gene pairs with significant angles and columns
are samples.
}
\description{
Constructs a matrix where each row is a pair of genes with a significant
angle and a column is a sample. Each cell in a matrix can be either 1 or
0, where 1 means that the significant angle between the pair of genes in
a row is present in this sample and 0 means not.
}
