% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_cons_nodes.R
\name{assemble_cons_nodes}
\alias{assemble_cons_nodes}
\title{Assebmle gene pairs with conserved angles}
\usage{
assemble_cons_nodes(l_processed, angle_type = c("sharp", "blunt"), fringe = 3)
}
\arguments{
\item{l_processed}{list. An output from the \strong{anglemanise} function.}

\item{angle_type}{character vector. Specifies which type of angles to
assemble. Can be "sharp", "blunt", or both.}

\item{fringe}{integer. Minimal number of samples where the relationship
between genes should be preserved, otherwise the gene pair is removed.
Default is 3.}
}
\value{
list. Each element contains a list with entries recording
significant angles per sample and an entry recording conserved angles
and their conservation level.
}
\description{
Extract unique gene features per dataset with significant
sharp/blunt angles.
}
