% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{batch_key}
\alias{batch_key}
\title{Access the Batch Key from an Anglem Object}
\usage{
batch_key(object)
}
\arguments{
\item{object}{An \code{anglem} object.}
}
\value{
A character string representing the batch key.
}
\description{
Retrieves the batch key used in the \code{anglem} object.
}
\examples{
\dontrun{
# Get the batch key
bkey <- batch_key(anglem_object)
}
}
