% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_anglemanise.R
\name{big_anglemanise}
\alias{big_anglemanise}
\title{Record critical angles between genes in gene expression
matrices from a Seurat list.}
\usage{
big_anglemanise(
  seurat_list,
  zscore_mean_threshold = 2,
  zscore_sn_threshold = 2,
  max_n_genes = 2000,
  n_cores = 4
)
}
\arguments{
\item{seurat_list}{seurat list of seurat objects with scaled data.}

\item{n_cores}{integer. Number of cores to use for
the computation of the cosine distances.}

\item{fdr_threshold}{double. The FDR threshold to apply
to the q-values. Defaults to 0.001.}
}
\value{
list. First two elements are sparse matrices
recording the number of sharp and blunt angles across
the datasets. Third and fourth elements are lists with
angles statistics and paths to the values of critical
angles.
}
\description{
\emph{anglemanise} is a high-level function that returns a list
of lists containing critical angles between genes across
input samples, angle statistics per sample, paths to the
the angle records per sample.
}
\details{
This is the main function of the package. It calculates
angles between all genes across all samples provided in
the Seurat list. It approximates the angle distribution
for each sample and extracts values of critical angles. Data needs to be scaled.
}
\seealso{
https://arxiv.org/abs/1306.0256
}
