% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_extract_corr.R
\name{big_extract_corr}
\alias{big_extract_corr}
\title{Calculate cosine angle between genes}
\usage{
big_extract_corr(x_mat, method = "pearson")
}
\arguments{
\item{x_mat}{An \code{\link[bigstatsr]{FBM}} object containing raw gene
expression data, where rows correspond to genes and columns to samples.
The data will be normalized and scaled within the function.}

\item{method}{A character string specifying the method to compute the
gene-gene relationships. Options are:
\itemize{
\item \code{"pearson"} (default): Computes the cosine angle between
genes.
\item \code{"spearman"}: Computes the Spearman rank correlation
coefficient by rank-transforming the data before computing the
correlation.
\item \code{"diem"}: Computes the Dimension Insensitive Euclidean
Metric between genes. Note that this is done in the
\code{\link{big_factorise}} function.
}}
}
\value{
An \code{\link[bigstatsr]{FBM}} object containing the gene-gene
correlation matrix. The matrix is square with dimensions equal to the
number of genes and contains the pairwise correlations between genes.
The diagonal elements are set to \code{NA}.
}
\description{
Constructs a matrix of gene-gene relationships based on distance
metrics.
}
\details{
The function returns the gene-gene angle matrix as an
\code{\link[bigstatsr]{FBM}} object.
}
\seealso{
\code{\link[bigstatsr]{big_apply}},
\code{\link[bigstatsr]{big_cor}},
\code{\link[bigstatsr]{FBM}},
\code{\link{big_factorise}}
}
