% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_factorise.R
\name{big_factorise}
\alias{big_factorise}
\title{Factorise angle matrices}
\usage{
big_factorise(x_mat, name)
}
\arguments{
\item{x_mat}{Matrix. Contains normalised and scaled gene
expression.}

\item{name}{character. The name of the dataset.}

\item{fdr_threshold}{double. Fraction of the correlation
to be cut from both sides of an approximated angle
distribution.}
}
\value{
list. First two elements are sparse matrices
containing the significant sharp and blunt angles
between genes. Third and fourth elements are lists
with angles statistics and paths to the values of
critical mangles.
}
\description{
Generates two a one-hot encoded square matrixces
recording the presence/absence (1/0) of sharp and blunt
critical angles.
}
\details{
\emph{factorise} extracts angles between genes, estimates
critical angles by approximation of the angle distribution,
and records angles passing the critical threshold into
a sparse matrix.
}
