% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglemanise_utils.R
\name{big_mat_list_mean}
\alias{big_mat_list_mean}
\title{Calculates the mean matrix of a list of FBMs.
Every element is summed together and divided by the number of FBMs.}
\usage{
big_mat_list_mean(fbmList)
}
\arguments{
\item{fbmList}{A list of FBM objects to be added.}
}
\value{
A new FBM object containing the sum of all FBMs in the list.
}
\description{
This function takes a list of FBMs, and calculates the mean of every element in the FBM.
If the list is empty or the FBMs have different dimensions, it throws an error.
}
\examples{
combined_matrix <- big_add_mat_list(list(fbm1, fbm2, fbm3))
}
