% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_unqiue_features.R
\name{count_unqiue_features}
\alias{count_unqiue_features}
\title{Count unique features per dataset}
\usage{
count_unqiue_features(conserved_nodes)
}
\arguments{
\item{conserved_nodes}{list. An output from the \strong{assemble_cons_nodes}
function.}
}
\value{
list. First element is a list where each entry carries a
a vector of unique features (gene names) under sepcific conservation
threshold. Second element is data.frame that records the number of
unique features per number of datasets intersected.
}
\description{
Counts and records unq9iue gene features per input dataset.
}
