% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{create_anglemaniaObject}
\alias{create_anglemaniaObject}
\title{Create an anglemaniaObject from a Seurat Object}
\usage{
create_anglemaniaObject(
  seurat_object,
  dataset_key = NA_character_,
  batch_key,
  min_cells_per_gene = 1
)
}
\arguments{
\item{seurat_object}{A \code{\link[Seurat]{Seurat}} object containing
single-cell RNA-seq data.}

\item{dataset_key}{A character string indicating the column name in the
Seurat object metadata that identifies the dataset to which each cell
belongs. If \code{NA}, all cells are assumed to belong to the same
dataset.}

\item{batch_key}{A character string indicating the column name(s) in the
Seurat object metadata that identify the batch to which each cell belongs.}

\item{min_cells_per_gene}{A numeric value indicating the minimum number of
cells in which a gene must be expressed to be included in the analysis.
Default is \code{1}.}
}
\value{
An \code{\link{anglemaniaObject}} containing:
\describe{
\item{\code{matrix_list}}{A list of filtered count matrices for each unique
batch.}
\item{\code{dataset_key}}{The dataset key used for splitting the Seurat
object.}
\item{\code{batch_key}}{The batch key used for splitting the Seurat
object.}
\item{\code{data_info}}{A data frame summarizing the number of samples per
dataset and their weights.}
\item{\code{weights}}{A numeric vector of weights for each dataset or batch
based on the number of samples.}
\item{\code{intersect_genes}}{A character vector of genes expressed in at
least the specified number of cells across all batches.}
\item{\code{min_cells_per_gene}}{The minimum number of cells per gene
threshold used for filtering.}
}
}
\description{
Constructs an \code{\link{anglemaniaObject}} from a given
\code{\link[Seurat]{Seurat}} object. This includes extracting and processing
count matrices, filtering genes based on expression in a minimum number of
cells, and storing results along with dataset and batch information. It also
calculates weights for each dataset or batch based on the number of samples.
}
\details{
The function performs the following steps:
\enumerate{
\item Adds a unique batch key to the Seurat object's metadata using
\code{\link{add_unique_batch_key}}.
\item Extracts count matrices for each batch.
\item Filters genes based on the \code{min_cells_per_gene} threshold.
\item Identifies intersected genes present across all batches.
\item Converts count matrices to \code{\link[bigstatsr]{FBM}} objects.
\item Computes weights for each batch or dataset.
}
}
\seealso{
\code{\link{anglemaniaObject}},
\code{\link{add_unique_batch_key}},
\code{\link{anglemania}},
\code{\link[bigstatsr]{FBM}}
}
