% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{data_info}
\alias{data_info}
\title{Access Data Information from an Anglem Object}
\usage{
data_info(object)
}
\arguments{
\item{object}{An \code{anglem} object.}
}
\value{
A data frame containing dataset and batch information.
}
\description{
Retrieves the data frame summarizing datasets and batches within the
\code{anglem} object.
}
\examples{
\dontrun{
# Get data information
info <- data_info(anglem_object)
}
}
