% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{dataset_key}
\alias{dataset_key}
\title{Access the Dataset Key from an Anglem Object}
\usage{
dataset_key(object)
}
\arguments{
\item{object}{An \code{anglem} object.}
}
\value{
A character string representing the dataset key.
}
\description{
Retrieves the dataset key used in the \code{anglem} object.
}
\examples{
\dontrun{
# Get the dataset key
dkey <- dataset_key(anglem_object)
}
}
