% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{.get_meta_data}
\alias{.get_meta_data}
\title{Internal helper function to get metadata from a Seurat or SCE object}
\usage{
.get_meta_data(object)
}
\description{
Internal helper function to get metadata from a Seurat or SCE object
}
\keyword{internal}
