% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_MOAV.R
\name{estimate_MOAV}
\alias{estimate_MOAV}
\title{Estiamte additional variance}
\usage{
estimate_MOAV(l_approx, s_dims)
}
\arguments{
\item{l_approx}{list. List with angular statistics after the
fitting of a gaussian.}

\item{s_dims}{integer. Vector with
probabilities corresponding to the angles in \strong{ref_data}.}
}
\value{
vector double. Vector of Kolmogorov-Smirnov D-statistics.
}
\description{
Estimates the measure of additional variance (MOAV) in reference
to the expected varience.
}
\details{
The function integrates non-overlapping area between the
analytical and theoritical distributions of angles between
vectors of gene expression.
}
