% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_angle_stats.R
\name{extract_angle_stats}
\alias{extract_angle_stats}
\title{Extract critical angles per dataset}
\usage{
extract_angle_stats(l_processed)
}
\arguments{
\item{l_processed}{list. An output from the \strong{anglemanise} function.}
}
\value{
data.frame. Records of critical angles per dataset.
}
\description{
Applied to the output of \strong{anglemanise} this function will record
critical angles estiamted per dataset into a data.frame.
}
