% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_integration_features.R
\name{extract_integration_features}
\alias{extract_integration_features}
\title{Extract integration features (gene names)}
\usage{
extract_integration_features(l_cumfact, cutoff = NULL)
}
\arguments{
\item{l_cumfact}{list. First element of the output from the
\strong{anglemanise} function.}

\item{cutoff}{integer. Number of datasets where the same relationship
between two genes is detected. Currently the simplest way to extract
integration: setting a brute cutoff on the level of conservation}
}
\value{
list. Two elements contain character vector with gene names
corresponding to sharp and blutn angles.
}
\description{
Number of gene pairs with conserved angles diminish along the number
of intersected dataset, seemingly resembling a Poisson distribution.
Here, we use the Poisson distribution to provide some sort of background
to the selection of the cutoff for the number of the intersected samples.
}
