% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{extract_integration_genes}
\alias{extract_integration_genes}
\title{Extract Integration Genes from an Anglem Object}
\usage{
extract_integration_genes(object)
}
\arguments{
\item{object}{An \code{anglem} object.}
}
\value{
A character vector of integration gene names.
}
\description{
Retrieves the list of genes selected for integration from the \code{anglem}
object.
}
\examples{
\dontrun{
# Extract integration genes
int_genes <- extract_integration_genes(anglem_object)
}
}
