% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurise.R
\name{featurise}
\alias{featurise}
\title{Extract gene featuries from the processed datasets}
\usage{
featurise(l_processed)
}
\arguments{
\item{l_processed}{list. An output from the \strong{anglemanise} function.}
}
\value{
list. List with unique gene features and conserved angles
between genes.
}
\description{
Extract unique gene features per dataset with significant
sharp/blunt angles.
}
