% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_angles.R
\name{filter_angles}
\alias{filter_angles}
\title{Filter a data.table with angles}
\usage{
filter_angles(x_df_ang, l_angles)
}
\arguments{
\item{x_df_ang}{data.table. A long data.table with three columns:
x - name of a gene in row, y - name of a gene in column,
angle - an angle between x and y.}

\item{l_angles}{list. An output of \strong{estimate_critical_angles}
function.}
}
\value{
list. List with two elements: data.tables with sharp
and blunt significant angles.
}
\description{
Filters out insignificant angles from an angle data.table.
}
\details{
This is a \strong{low-level} function that utilises functionality
of \strong{data.table} package to filter insgificant entries from
an angle data.table.
}
