% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gaussian.R
\name{fit_gaussian}
\alias{fit_gaussian}
\title{Fit a guassian distribution}
\usage{
fit_gaussian(l_approx)
}
\arguments{
\item{l_approx}{list. An output from the \strong{approxinate.angles} function.}
}
\value{
list. List with multiple entries, one of which contains
an approxiamted distribution and details of the gaussian fit.
Others contains NAs to be filled by subsequent functions.
}
\description{
Fits the gaussian distribution to the approximated angle distribution
using the sum of residual squares.
}
\details{
Fitting is done using non-linear minimization from the \strong{stats} package.
}
