% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flippitise.R
\name{flippitise}
\alias{flippitise}
\title{Estimate critical angles}
\usage{
flippitise(x_mat, x_df_ang, l_angles)
}
\arguments{
\item{x_df_ang}{data.table. A long data.table with three columns:
x - name of a gene in row, y - name of a gene in column,
angle - an angle between x and y.}

\item{s_dims}{integer. Number of samples in the input gene
expression matrix.}

\item{extrema}{double. Fraction of the angles
to be cut from both sides of an approximated angle
distribution.}
}
\value{
list. Records on parameters of the gaussian,
goodness of fit, and critical angles.
}
\description{
A wrapper around functions that estimate_critical_angles
and their levels of signficance from an approximated
distribution of angles.
}
\details{
First, the angle distribution is approximated by quantile
splitting providid by the user. Then the gaussian is fit by
quantile transformation. Goodness of fit is tested by
Kolmogorov-Smirnov test. Measure of additional variance is
estiated fron the data by integrating the difference between
the fitted gaussian and theoretical distribution. Critical
angles are extracted from the fitted gaussian.
}
