% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analytical_extremes.R
\name{get_analytical_extremes}
\alias{get_analytical_extremes}
\title{Test goodness of fit}
\usage{
get_analytical_extremes(l_approx, extrema)
}
\arguments{
\item{l_approx}{list. An output from the \strong{approxinate.angles} function.}

\item{alpha}{double. Significance level for the goodness of fit.}

\item{mt_nsims}{integer. Number of simulated Ds.}

\item{seed}{integer. Random seed. Important for reproducibility
of the background simulation.}
}
\value{
list. List with multiple entries, one of which contains
an approxiamted distribution and details of the gaussian fit.
Others contains NAs to be filled by subsequent functions.
}
\description{
Test the goodness of fit of the gaussian with the Kolmogorov-Smirnov
test.
}
\details{
Goodness of fit is determined by comparing the analztical D against
the simulation background distribution of D. One-sided p-value is
computed agaisnt a provided alpha.
}
