% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_critangs_paths.R
\name{get_critangs_paths}
\alias{get_critangs_paths}
\title{Get paths to files with critical angles}
\usage{
get_critangs_paths(l_processed, samp_id, angle_type = c("sharp", "blunt"))
}
\arguments{
\item{l_processed}{list. An output from the \strong{anglemanise} function.}

\item{samp_id}{character vector. Samples name or names to look for.}

\item{angle_type}{character vector. Specifies which type of angles to
look for. Can be "sharp", "blunt", or both.}
}
\value{
character vector. Vector of paths to the files where critical
angles were recorded.
}
\description{
Collects paths to the files where the critical angles were recorded to
during the processing of the input data by the \strong{anglemanise} function.
}
