% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglemanise_utils.R
\name{get_dstat}
\alias{get_dstat}
\title{Compute mean and standard deviation of the correlation matrix}
\usage{
get_dstat(corr_matrix)
}
\arguments{
\item{corr_matrix}{A FBM object from the bigstatsr package.}
}
\value{
A list with two entries: \code{mean} and \code{sd}.
}
\description{
Computes the mean and standard deviation of the correlation matrix
using the big_apply function.
}
