% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plottab.R
\name{get_plottab}
\alias{get_plottab}
\title{Construct a plottiugn table from a seurat object}
\usage{
get_plottab(seurat_object, var, reduction)
}
\arguments{
\item{seurat_object}{seurat object.}

\item{var}{character vector. A vector of varaibles in meta.data slot
to append.}

\item{reduction}{character. Name of the reduction method to pull
embeddings from.}
}
\value{
data.frame.
}
\description{
Exctracts a data.frame with embeddings for a specified reduction method
and appends specified variables from the meta.data slot.
}
