% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sl.R
\name{import_sl}
\alias{import_sl}
\title{Read in Seacells from a .RDS file}
\usage{
import_sl(path_dat = NULL, min_mcells = 15)
}
\arguments{
\item{path_dat}{string. Must be a path to an existing
.RDS file with a SeuratList of samples.}

\item{min_mcells}{integer. Minimal number of Metacells
ppresent in a sample. If less, a sample is discarded.
Defaulit is 15.}
}
\value{
Seurat list.
}
\description{
import_sl returns a Seurat list with normalised and
scaled expression matrices.
}
\details{
This function doesn't assess if the read object is a
Seurat object or not, and therefore relies on the user
to provide the path to correct data. Although the function will
read and pre-process any Seurat list, metacells are expected as
an input.
}
