% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_by_features.R
\name{integrate_seurat_list}
\alias{integrate_seurat_list}
\title{Integrate a List of Seurat Objects Using Selected Features}
\usage{
integrate_seurat_list(
  seurat_list,
  features,
  int_order = NULL,
  process = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{seurat_list}{A list of \code{\link[Seurat]{Seurat}} objects to be
integrated.}

\item{features}{A character vector of gene names (features) used for
integration.}

\item{int_order}{An optional data frame specifying the integration order
of samples within the Seurat list. See the \code{sample.tree} argument
in \code{\link[Seurat]{IntegrateData}} for more details. If not
provided, Seurat will construct the integration order using hierarchical
clustering. Default is \code{NULL}.}

\item{process}{Logical value indicating whether to further process the
data after integration (i.e., scale it, run PCA, and compute UMAP
embeddings). Default is \code{TRUE}.}

\item{verbose}{Logical value indicating whether to display progress
messages during integration. Default is \code{FALSE}.}
}
\value{
A \code{\link[Seurat]{Seurat}} object containing the integrated
data. The default assay is set to \code{"integrated"}.
}
\description{
\code{integrate_seurat_list} integrates a list of Seurat objects (e.g.,
representing different samples or batches) using canonical correlation
analysis (CCA) based on a set of selected features (genes). The function
handles normalization, finding integration anchors, integrating data, and
optional downstream processing steps such as scaling, PCA, and UMAP
visualization.
}
\details{
The function performs the following steps:
\enumerate{
\item \strong{Normalization}: Each Seurat object in the list is log-normalized
using \code{\link[Seurat]{NormalizeData}}.
\item \strong{Parameter Adjustment}: Integration parameters are adjusted based
on the smallest dataset to accommodate cases with a small number of
cells (e.g., metacells).
\item \strong{Finding Integration Anchors}: Uses
\code{\link[Seurat]{FindIntegrationAnchors}} to find anchors between
datasets based on the provided features.
\item \strong{Integration}: Integrates the datasets using
\code{\link[Seurat]{IntegrateData}}.
\item \strong{Optional Processing}: If \code{process = TRUE}, the function
scales the data, runs PCA, and computes UMAP embeddings.
}

The integration is performed using Seurat's CCA-based methods, and the
function is designed to handle datasets with varying sizes efficiently.
}
\examples{
\donttest{
# Integrate a list of seurat object using selected 
#  features (e.g. anglemania genes or HVGs)
#  and CCA integration method
seurat_list <- list(seurat_object1, seurat_object2)
integrated_seurat <- integrate_seurat_list(seurat_list, features)
}
}
\seealso{
\code{\link{integrate_by_features}},
\code{\link[Seurat]{IntegrateData}},
\code{\link[Seurat]{FindIntegrationAnchors}},
\code{\link[Seurat]{NormalizeData}},
\code{\link[Seurat]{ScaleData}},
\code{\link[Seurat]{RunPCA}},
\code{\link[Seurat]{RunUMAP}}
}
