% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{intersect_genes<-}
\alias{intersect_genes<-}
\title{Set Intersected Genes in an Anglem Object}
\usage{
intersect_genes(object) <- value
}
\arguments{
\item{object}{An \code{anglem} object.}

\item{value}{A character vector of gene names.}
}
\value{
The updated \code{anglem} object.
}
\description{
Assigns a new vector of intersected genes to the \code{anglem} object.
}
\examples{
\dontrun{
# Set new intersected genes
intersect_genes(anglem_object) <- new_gene_list
}
}
