% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{intersect_genes}
\alias{intersect_genes}
\title{Access Intersected Genes from an Anglem Object}
\usage{
intersect_genes(object)
}
\arguments{
\item{object}{An \code{anglem} object.}
}
\value{
A character vector of intersected gene names.
}
\description{
Retrieves the vector of genes that are expressed in at least the specified
number of cells across all batches.
}
\examples{
\dontrun{
# Get intersected genes
genes <- intersect_genes(anglem_object)
}
}
