% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{list_stats<-}
\alias{list_stats<-}
\title{Set Statistical Measures in an Anglem Object}
\usage{
list_stats(object) <- value
}
\arguments{
\item{object}{An \code{anglem} object.}

\item{value}{A list containing statistical matrices.}
}
\value{
The updated \code{anglem} object.
}
\description{
Assigns a new list of statistical measures to the \code{anglem} object.
}
\examples{
\dontrun{
# Set new list statistics
list_stats(anglem_object) <- new_stats_list
}
}
