% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{list_stats}
\alias{list_stats}
\title{Access Statistical Measures from an Anglem Object}
\usage{
list_stats(object)
}
\arguments{
\item{object}{An \code{anglem} object.}
}
\value{
A list containing statistical matrices such as mean z-scores and SNR
z-scores.
}
\description{
Retrieves the list of statistical measures computed across datasets in the
\code{anglem} object.
}
\examples{
\dontrun{
# Get list statistics
stats <- list_stats(anglem_object)
}
}
