% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{matrix_list<-}
\alias{matrix_list<-}
\title{Set the Matrix List in an Anglem Object}
\usage{
matrix_list(object) <- value
}
\arguments{
\item{object}{An \code{anglem} object.}

\item{value}{A list of \code{\link[bigstatsr]{FBM}} objects.}
}
\value{
The updated \code{anglem} object.
}
\description{
Assigns a new list of gene expression matrices to the \code{anglem} object.
}
\examples{
\dontrun{
# Set a new matrix list
matrix_list(anglem_object) <- new_matrix_list
}
}
