% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{matrix_list}
\alias{matrix_list}
\title{Access the Matrix List from an Anglem Object}
\usage{
matrix_list(object)
}
\arguments{
\item{object}{An \code{anglem} object.}
}
\value{
A list of \code{\link[bigstatsr]{FBM}} objects containing gene
expression matrices.
}
\description{
Retrieves the list of gene expression matrices stored in the \code{anglem}
object.
}
\examples{
\dontrun{
# Access the matrix list
matrices <- matrix_list(anglem_object)
}
}
