% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_to_df.R
\name{melt_to_df}
\alias{melt_to_df}
\title{Melt distance matrix}
\usage{
melt_to_df(x_mat_ang)
}
\arguments{
\item{x_mat_ang}{matrix. Contains angles between the
vectors of gene expression between genes.}
}
\value{
data.table. A long data.table with three columns:
x - name of a gene in row, y - name of a gene in column,
angle - an angle between x and y.
}
\description{
Melts an angle matrix into a long data.frame.
}
\details{
This is a \strong{low-level} function that utilises functionality
of \strong{data.table} pacckage to melt a square angle matrix into
a long data.frame.
}
