% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlo_Dstat.R
\name{montecarlo_Dstat}
\alias{montecarlo_Dstat}
\title{Simulate background D}
\usage{
montecarlo_Dstat(ref_data, ref_quants, mu, sigma, nsamp = 2000, seed = 42)
}
\arguments{
\item{ref_data}{vector double. Vector with angles
corresponding to the probabilities in \strong{ref_quants}.}

\item{ref_quants}{vector double. Vector with
probabilities corresponding to the angles in \strong{ref_data}.}

\item{mu}{double. Mean of the fitted gaussian.}

\item{sigma}{double. Standard deviation of the
fitted gaussian.}

\item{nsamp}{integer. Number of the simulated distributions.}

\item{seed}{integer. Random seed. Important for reproducibility
of the background simulation.}
}
\value{
vector double. Vector of Kolmogorov-Smirnov D-statistics.
}
\description{
Monte-Carlo simulation of background D statistic distribution from
estimated parameters of a Gaussian distribution
}
\details{
The function samples \strong{nsamp} times from a gaussian distribution
with mean and standard deviation provided by the user. Then, the
samples are tested against the provided reference and resulting
D-statistics are recorded in the output vector.
}
