% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglemanise_utils.R
\name{permute_counts}
\alias{permute_counts}
\title{Permute counts in a Seurat object}
\usage{
permute_counts(X.sub, which = "columns", seed = 1)
}
\arguments{
\item{which}{A character string specifying whether to permute "columns" or "rows".}

\item{seu}{A Seurat object.}
}
\value{
A new Seurat object with permuted count data.
}
\description{
This function permutes the counts in the assay data of a Seurat object either by columns or rows.
}
\examples{
seu_permuted = permute_counts(seu, which = "columns")
}
