% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_angles.R
\name{permute_nonzero}
\alias{permute_nonzero}
\title{permute non-zero elements of vectors}
\usage{
permute_nonzero(v)
}
\arguments{
\item{v}{A numeric vector.}
}
\value{
A numeric vector with non-zero elements permuted.
}
\description{
Permutes the non-zero elements of a numeric vector.
}
\examples{
v <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10) # put in some zeroes
v = c(v, 0, 0, 0)
permute_nonzero(v)
}
