% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglemanise_utils.R
\name{prefilter_angl}
\alias{prefilter_angl}
\title{Preselect Genes from an anglemania_object, to make subsequent filtering easier}
\usage{
prefilter_angl(angl, zscore_mean_threshold = 1, zscore_sn_threshold = 1)
}
\arguments{
\item{angl}{An \code{anglemania_object} containing statistical
matrices such as mean z-scores and SNR z-scores.}

\item{zscore_mean_threshold}{Numeric value specifying the threshold for the
absolute mean z-score. Default is 1.}

\item{zscore_sn_threshold}{Numeric value specifying the threshold for the
SNR z-score. Default is 1.}
}
\value{
The input \code{anglemania_object} with the
\code{integration_genes} slot updated to include the selected genes and
their statistical information.
}
\description{
Preselect Genes from an anglemania_object, to make subsequent filtering easier
}
\details{
The function performs the following steps:
\enumerate{
\item Checks if the input object is of class \code{\link{anglemania_object-class}}.
\item Identifies gene pairs where both the mean z-score and SNR z-score
exceed the specified thresholds.
}
}
\keyword{internal}
