% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{sce_example}
\alias{sce_example}
\title{Generate example SingleCellExperiment object}
\usage{
sce_example(seed = 42)
}
\arguments{
\item{seed}{The seed for the random number generator. Because this function
is only used locally, we allow to set a seed within the function.}
}
\value{
A SingleCellExperiment object
}
\description{
This function generates a SingleCellExperiment object with 2 batches and 2
datasets. The object contains 300 genes and 600 cells. The counts matrix is
generated using the \code{rpois} function with different lambda values for the
two batches.
}
\examples{
sce <- sce_example()
sce
}
