% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglemanise_utils.R
\name{sparse_to_fbm}
\alias{sparse_to_fbm}
\title{Convert a sparse matrix into an FBM}
\usage{
sparse_to_fbm(s_mat)
}
\arguments{
\item{s_mat}{A sparse matrix.}
}
\value{
A FBM object from the bigstatsr package.
}
\description{
Converts a sparse matrix into an FBM with efficient memory usage.
}
