% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_gof_Ks.R
\name{test_gof_Ks}
\alias{test_gof_Ks}
\title{Test goodness of fit}
\usage{
test_gof_Ks(l_approx, alpha = 0.01, mt_nsims = 6000, seed = 42)
}
\arguments{
\item{l_approx}{list. An output from the \strong{approxinate.angles} function.}

\item{alpha}{double. Significance level for the goodness of fit.}

\item{mt_nsims}{integer. The number of simulated distributions.}

\item{seed}{integer. Random seed. Important for reproducibility
of the background simulation.}
}
\value{
list. List with multiple entries, one of which contains
an approxiamted distribution and details of the gaussian fit.
Others contains NAs to be filled by subsequent functions.
}
\description{
Kolmogorov-Smirnov test for the goodness of fit.
}
\details{
Goodness of fit is determined by comparing the analztical D against
the simulation background distribution of D. One-sided p-value is
computed against a provided alpha.
}
