% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{weights<-}
\alias{weights<-}
\title{Set Weights in an Anglem Object}
\usage{
weights(object) <- value
}
\arguments{
\item{object}{An \code{anglem} object.}

\item{value}{A named numeric vector of weights.}
}
\value{
The updated \code{anglem} object.
}
\description{
Assigns new weights to the datasets or batches in the \code{anglem} object.
}
\examples{
\dontrun{
# Set new weights
weights(anglem_object) <- c(batch1 = 0.5, batch2 = 0.5)
}
}
