% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{weights}
\alias{weights}
\title{Access Weights from an Anglem Object}
\usage{
weights(object, ...)
}
\arguments{
\item{object}{An \code{anglem} object.}
}
\value{
A named numeric vector of weights.
}
\description{
Retrieves the weights assigned to each dataset or batch in the \code{anglem}
object.
}
\examples{
\dontrun{
# Get weights
wts <- weights(anglem_object)
}
}
