% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_angles.R
\name{write_angles}
\alias{write_angles}
\title{Write angles as tab-delimted tables}
\usage{
write_angles(l_critical, path_to_write_angles)
}
\arguments{
\item{l_critical}{list. List with two elements: data.tables with sharp
and blunt significant angles.}

\item{path_to_write_angles}{string. Path to a directory to write tables.}
}
\value{
list. List with md5-sums correspondign to the two written tables.
}
\description{
Writes to the disk two tables with angles in a tab-delimited format.
}
\details{
This is a \strong{low-level} function that uses \strong{data.table's} package
fwrite function to write tables with croitical angles to the disk.
Names of the written tables are md5-sums of the same tables.
}
