% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglemania-package.R
\docType{package}
\name{anglemania-package}
\alias{anglemania}
\alias{anglemania-package}
\title{anglemania: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\section{anglemania functions}{

The anglemania functions ...
}

\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
