% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglemania.R
\name{anglemania}
\alias{anglemania}
\title{Compute Critical Angles Between Genes Across Samples in an anglemaniaObject}
\usage{
anglemania(
  anglemania_object,
  method = "pearson",
  zscore_mean_threshold = 2.5,
  zscore_sn_threshold = 2.5,
  max_n_genes = 2000
)
}
\arguments{
\item{anglemania_object}{An \code{\link{anglemaniaObject}} containing gene
expression data and associated metadata.}

\item{method}{Character string specifying the method to use for calculating
the relationship between gene pairs. Default is \code{"pearson"}.
Other options include \code{"diem"}
(see \url{https://bytez.com/docs/arxiv/2407.08623/paper}).}

\item{zscore_mean_threshold}{Numeric value specifying the threshold
for the mean z-score. Default is \code{2.5}.}

\item{zscore_sn_threshold}{Numeric value specifying the threshold for the
signal-to-noise z-score. Default is \code{2.5}.}

\item{max_n_genes}{Integer specifying the maximum number of genes to select.
Default is \code{2000}.}
}
\value{
An updated \code{\link{anglemaniaObject-class}} with computed statistics and
selected genes based on the specified thresholds.
}
\description{
\code{anglemania} computes critical angles between genes across all samples
provided in an \code{\link{anglemaniaObject}}. It calculates angles,
transforms them to z-scores, computes statistical measures, and selects
genes based on specified thresholds.
}
\details{
This function performs the following steps:
\enumerate{
\item Computes angles between genes for each sample in the
\code{anglemania_object} using the specified \code{method}, via
\code{\link{factorise}}.
\item Transforms the angles to z-scores.
\item Computes statistical measures (mean z-score, signal-to-noise ratio)
across samples using \code{\link{get_list_stats}}.
\item Selects genes based on specified z-score thresholds using
\code{\link{select_genes}}.
}

The computed statistics and selected genes are added to the
\code{anglemania_object}, which is returned.
}
\examples{
load(system.file(
 "extdata",
 "seurat_splatter_sim.RData",
 package = "anglemania"))
batch_key <- "Batch"
anglemania_object <- create_anglemaniaObject(se,
 batch_key = batch_key,
 min_cells_per_gene = 1
 )

anglemania_object <- anglemania(
  anglemania_object,
  method = "pearson",
  zscore_mean_threshold = 2,
  zscore_sn_threshold = 2,
  max_n_genes = 2000
)

# Access the selected genes
selected_genes <- get_anglemania_genes(anglemania_object)
selected_genes[1:10]
}
\seealso{
\code{\link{create_anglemaniaObject}},
\code{\link{get_list_stats}},
\code{\link{select_genes}},
\code{\link{factorise}},
\code{\link[bigstatsr]{big_apply}},
\url{https://arxiv.org/abs/1306.0256}
}
