% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglemania.R
\name{anglemania}
\alias{anglemania}
\title{Compute Critical Angles Between Genes Across Samples in an anglemania_object}
\usage{
anglemania(
  angl,
  zscore_mean_threshold = 2.5,
  zscore_sn_threshold = 2.5,
  max_n_genes = NULL,
  method = "cosine",
  permute_row_or_column = "column",
  permutation_function = "sample",
  prefilter_threshold = 0.5
)
}
\arguments{
\item{angl}{An \code{\link{anglemania_object-class}} containing gene
expression data and associated metadata.}

\item{zscore_mean_threshold}{Numeric value specifying the threshold
for the mean z-score. Default is \code{2.5}.}

\item{zscore_sn_threshold}{Numeric value specifying the threshold for the
signal-to-noise z-score. Default is \code{2.5}.}

\item{max_n_genes}{Integer specifying the maximum number of genes to select.}

\item{method}{Character string specifying the method to use for calculating
the relationship between gene pairs. Default is \code{"cosine"}.
Other options include \code{"diem"}
(see \url{https://bytez.com/docs/arxiv/2407.08623/paper}).}

\item{permute_row_or_column}{Character "row" or "column", whether permutations should be executed row-wise or column wise. Default is \code{"column"}}

\item{permutation_function}{Character "sample" or "permute_nonzero". If sample, then sample is used for constructing background distributions. If permute_nonzero, then only non-zero values are permuted. Default is \code{"sample"}}

\item{prefilter_threshold}{Numeric value specifying the threshold prefiltering genes - speeds up gene selection}
}
\value{
An updated \code{\link{anglemania_object-class}} with computed statistics and
selected genes based on the specified thresholds.
}
\description{
\code{anglemania} computes critical angles between genes across all samples
provided in an \code{\link{anglemania_object-class}}. It calculates angles,
transforms them to z-scores, computes statistical measures, and selects
genes based on specified thresholds.
}
\details{
This function performs the following steps:
\enumerate{
\item Computes angles between genes for each sample in the
\code{angl} using the specified \code{method}, via
\code{\link{factorise}}.
\item Transforms the angles to z-scores.
\item Computes statistical measures (mean z-score, signal-to-noise ratio)
across samples using \code{\link{get_list_stats}}.
\item Selects genes based on specified z-score thresholds using
\code{\link{select_genes}}.
}

The computed statistics and selected genes are added to the
\code{angl}, which is returned.
}
\examples{
# Set seed for reproducibility (optional)
sce <- sce_example()
angl <- create_anglemania_object(sce, batch_key = "batch")
angl <- anglemania(
  angl,
  method = "cosine",
  zscore_mean_threshold = 2,
  zscore_sn_threshold = 2,
  max_n_genes = 2000
)

# Access the selected genes
selected_genes <- get_anglemania_genes(angl)
selected_genes[1:10]
}
\seealso{
\code{\link{create_anglemania_object}},
\code{\link{get_list_stats}},
\code{\link{select_genes}},
\code{\link{factorise}},
\code{\link[bigstatsr]{big_apply}},
\url{https://arxiv.org/abs/1306.0256}
}
