% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglemanise_utils.R
\name{big_mat_list_mean}
\alias{big_mat_list_mean}
\title{Calculates the mean for each element of a matrix from a list of FBMs}
\usage{
big_mat_list_mean(anglemania_object)
}
\arguments{
\item{anglemania_object}{An \code{anglemaniaObject} containing the list of FBMs.
In this case, the FBMs are the angle matrices computed in \code{factorise}.}
}
\value{
A new \code{\link[bigstatsr]{FBM}} object containing the mean values.
}
\description{
This function takes an \code{anglemaniaObject} containing a list of
\code{\link[bigstatsr]{FBM}}s and calculates the mean of every element.
If the list is empty or the FBMs have different dimensions,
it throws an error.
}
\examples{
# Create FBMs
mat1 <- matrix(1:9, nrow = 3)
mat2 <- matrix(1:9, nrow = 3)

fbm1 <- bigstatsr::FBM(nrow = nrow(mat1), ncol = ncol(mat1), init = mat1)
fbm2 <- bigstatsr::FBM(nrow = nrow(mat2), ncol = ncol(mat2), init = mat2)

# Create weights
weights <- c(batch1 = 0.5, batch2 = 0.5)

# Create the list of FBMs
fbm_list <- list(batch1 = fbm1, batch2 = fbm2)

# Construct the anglemaniaObject
anglemania_object <- new(
  "anglemaniaObject",
  weights = weights,
  matrix_list = fbm_list
)
big_mat_list_mean(anglemania_object)
}
