% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglemanise_utils.R
\name{big_mat_list_mean}
\alias{big_mat_list_mean}
\title{Calculates the mean for each element of a matrix from a list of FBMs}
\usage{
big_mat_list_mean(anglemania_object)
}
\arguments{
\item{anglemania_object}{An \code{anglemaniaObject} containing the list of FBMs.}
}
\value{
A new \code{\link[bigstatsr]{FBM}} object containing the mean values.
}
\description{
This function takes an \code{anglemaniaObject} containing a list of
\code{\link[bigstatsr]{FBM}}s and calculates the mean of every element.
If the list is empty or the FBMs have different dimensions,
it throws an error.
}
