% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_angles.R
\name{extract_angles}
\alias{extract_angles}
\title{Construct a matrix of angles between genes}
\usage{
extract_angles(x_mat, n_threads = 16)
}
\arguments{
\item{x_mat}{matrix. Contains normalised and scaled gene
expression, where rows are genes and columns are samples.}

\item{n_threads}{integer. Number of cores to use for
the computation of the cosine distances.}
}
\value{
matrix. Square matrix containing angles between
vectors of gene expression (rwos of the input matrix).
}
\description{
Constructs a matrix of angles between genes from an input
gene expression matrix.
}
\details{
This is a \strong{low-level} function that utilises functionality
of \strong{parallelDist} pacakge to compute cosine distances
between vectors of gene expression. Since \strong{parallelDist}´s#
outputs a dist object, an Rcpp functions is used to convert it
back to matrix.
}
