% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorise.R
\name{factorise}
\alias{factorise}
\title{Factorise angle matrices}
\usage{
factorise(x_mat, extrema, n_threads, path_to_write_angles)
}
\arguments{
\item{x_mat}{Matrix. Contains normalised and scaled gene
expression.}

\item{extrema}{double. Fraction of the angles
to be cut from both sides of an approximated angle
distribution.}

\item{n_threads}{integer. Number of cores to use for
the computation of the cosine distances.}

\item{path_to_write_angles}{string. A path, under which
a temporary directory will be created to record values
of critical angles for each sample.}
}
\value{
list. First two elements are sparse matrices
containing the significant sharp and blunt angles
between genes. Third and fourth elements are lists
with angles statistics and paths to the values of
critical mangles.
}
\description{
Generates two a one-hot encoded square matrixces
recording the presence/absence (1/0) of sharp and blunt
critical angles.
}
\details{
\emph{factorise} extracts angles between genes, estimates
critical angles by approximation of the angle distribution,
and records angles passing the critical threshold into
a sparse matrix.
}
