% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglemanise_utils.R
\name{get_dstat}
\alias{get_dstat}
\title{Compute mean and standard deviation of the correlation matrix}
\usage{
get_dstat(corr_matrix)
}
\arguments{
\item{corr_matrix}{An \code{\link[bigstatsr]{FBM}} object.}
}
\value{
A list with statistical measures including \code{mean}, \code{sd},
\code{var}, \code{sn}, \code{min}, and \code{max}.
}
\description{
Computes the mean and standard deviation of the correlation matrix using the
\code{big_apply} function.
}
\examples{
s_mat <- Matrix::rsparsematrix(nrow = 10, ncol = 5, density = 0.3)
# Convert the sparse matrix to an FBM using your function
fbm_mat <- sparse_to_fbm(s_mat)
result <- get_dstat(fbm_mat)
str(result)
result
}
\seealso{
\code{\link[bigstatsr]{big_apply}}, \code{\link[bigstatsr]{FBM}}
}
