% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglemanise_utils.R
\name{get_list_stats}
\alias{get_list_stats}
\title{Calculate statistical measures from a list of FBMs}
\usage{
get_list_stats(anglemania_object)
}
\arguments{
\item{anglemania_object}{An \code{anglemaniaObject} containing the list of FBMs.}
}
\value{
A list containing three matrices: \code{mean_zscore},
\code{sds_zscore}, and \code{sn_zscore}.
}
\description{
Computes the mean, standard deviations, and signal-to-noise ratio (SNR)
for each element across a list of FBMs in an \code{anglemaniaObject}.
}
\examples{
load(system.file(
"extdata",
 "seurat_splatter_sim.RData",
 package = "anglemania")
)
anglemania_object <- create_anglemaniaObject(se, batch_key = "Batch")
anglemania_object <- anglemania(anglemania_object)
list_stats(anglemania_object) <- get_list_stats(anglemania_object)
str(list_stats(anglemania_object))
}
\seealso{
\code{\link[bigstatsr]{big_apply}}, \code{\link[bigstatsr]{FBM}}
}
