% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglemanise_utils.R
\name{get_list_stats}
\alias{get_list_stats}
\title{Calculate statistical measures from a list of FBMs}
\usage{
get_list_stats(angl)
}
\arguments{
\item{angl}{An \code{anglemania_object} containing the list of FBMs.}
}
\value{
A list containing three matrices: \code{mean_zscore},
\code{sds_zscore}, and \code{sn_zscore}.
}
\description{
Computes the mean, standard deviations, and signal-to-noise ratio (SNR)
for each element across a list of FBMs in an \code{anglemania_object}.
}
\examples{
se <- SeuratObject::pbmc_small
angl <- create_anglemania_object(se, batch_key = "groups")
angl <- anglemania(angl)
list_stats(angl) <- get_list_stats(angl)
str(list_stats(angl))
}
\seealso{
\code{\link[bigstatsr]{big_apply}}, \code{\link[bigstatsr]{FBM}}
}
