% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sl.R
\name{import_sl}
\alias{import_sl}
\title{Read in Seacells from a .RDS file}
\usage{
import_sl(seurat_list, min_mcells = 6)
}
\arguments{
\item{seurat_list}{Seurat list.}

\item{min_mcells}{integer. Minimal number of Metacells
present in a sample. If less, a sample is discarded.
Defaulit is 6. Integration using Seurat CCA seemingly
doesn't work with <6 cells.}
}
\value{
Seurat list.
}
\description{
import_sl takes a list of Seurat objects, checks if the conditions for
anglemanise are correct.
Briefly, it checks
a) if all objects have more than 'mcells'. Discard those samples with less.
b) a layer with scaled data. If not, it performs SCTransform (using glampoi)
returns a list of corrected Seurat objects
}
\details{
This function doesn't assess if the read object is a
Seurat object or not, and therefore relies on the user
to provide the path to correct data. Although the function will
read and pre-process any Seurat list, metacells are expected as
an input.
}
