% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_by_features.R
\name{integrate_by_features}
\alias{integrate_by_features}
\title{Integrate samples in a seurat list}
\usage{
integrate_by_features(seurat_list, features_to_integrate, int_order = NULL)
}
\arguments{
\item{seurat_list}{seurat list.}

\item{features_to_integrate}{character vector. Vector of gene names
(features) to construct anchors from.}

\item{int_order}{data.frame. Table specifying the integration order
of samples within the seurat list. See Seurat::IntegrateData's argument
sample.tree help page for more details. If not provided Seurat will
construct the integration order from hclust. Default is NULL.}
}
\value{
seurat object. Integrated seurat object. Seurat::DefaultAssay
is set to "integrated".
}
\description{
Integrates samples in the seurat list using CCA and parameters devised
by VF.
}
