% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_by_features.R
\name{integrate_by_features}
\alias{integrate_by_features}
\title{Integrate Samples in a Seurat Object Using Selected Features from
\code{anglem_object} Object}
\usage{
integrate_by_features(
  seurat_object,
  anglem_object,
  int_order = NULL,
  process = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{seurat_object}{A \code{\link[Seurat]{Seurat}} object containing
all samples or batches to be integrated.}

\item{anglem_object}{An \code{\link{anglem}} object previously generated
using \code{\link{create_anglem}} and \code{\link{big_anglemanise}}.
It is important that the \code{dataset_key} and \code{batch_key} are
correctly set in the \code{anglem_object}.}

\item{int_order}{An optional data frame specifying the integration order
of samples within the Seurat list. See the \code{sample.tree} argument
in \code{\link[Seurat]{IntegrateData}} for more details. If not
provided, Seurat will construct the integration order using
hierarchical clustering. Default is \code{NULL}.}

\item{process}{Logical value indicating whether to further process the
data after integration (i.e., scale it, run PCA, and compute UMAP
embeddings). Default is \code{TRUE}.}

\item{verbose}{Logical value indicating whether to display progress
messages during integration. Default is \code{FALSE}.}
}
\value{
A \code{\link[Seurat]{Seurat}} object containing the integrated
data. The default assay is set to \code{"integrated"}.
}
\description{
\code{integrate_by_features} integrates samples or batches within a Seurat
object using canonical correlation analysis (CCA) based on a set of
selected features (genes). The function utilizes an \code{anglem} object to
extract integration genes and handles the integration process, including
optional downstream processing steps such as scaling, PCA, and UMAP
visualization.
}
\details{
The function performs the following steps:
\enumerate{
\item \strong{Batch Key Addition}: Adds a unique batch key to the Seurat
object's metadata to distinguish different batches or samples. Batch
key is set to the \code{anglem_object}'s \code{batch_key}.
\item \strong{Splitting}: Splits the Seurat object into a list of Seurat
objects based on the batch key.
\item \strong{Integration}: Calls \code{\link{integrate_seurat_list}} to
integrate the list of Seurat objects using the features extracted
from the \code{anglem_object}.
}

The integration is performed using Seurat's CCA-based methods, and
parameters are adjusted based on the smallest dataset to ensure
compatibility with small sample sizes (e.g., metacells or SEACells). If
\code{process = TRUE}, the function will also scale the data, run PCA,
and compute UMAP embeddings.
}
\examples{
\dontrun{
# Load required libraries
library(Seurat)
library(pbapply)

# Assume seurat_object is a Seurat object containing all samples/batches
# and anglem_object is an anglem object with integration genes identified

# Integrate the Seurat object using the anglem object
integrated_seurat <- integrate_by_features(
  seurat_object = seurat_object,
  anglem_object = anglem_object,
  process = TRUE,
  verbose = TRUE
)

# Access the integrated assay
DefaultAssay(integrated_seurat) # Should be "integrated"

# Visualize UMAP embedding
DimPlot(integrated_seurat, reduction = "umap", group.by = "batch")
}

}
\seealso{
\code{\link{create_anglem}},
\code{\link{big_anglemanise}},
\code{\link{extract_integration_genes}},
\code{\link{integrate_seurat_list}},
\code{\link[Seurat]{IntegrateData}},
\code{\link[Seurat]{FindIntegrationAnchors}}
}
