% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_by_features.R
\name{integrate_by_features}
\alias{integrate_by_features}
\title{Integrate Samples in a Seurat Object Using Selected Features from
\code{anglemania_object} Object}
\usage{
integrate_by_features(
  seurat_object,
  angl,
  int_order = NULL,
  process = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{seurat_object}{A \code{\link[Seurat]{Seurat}} object containing
all samples or batches to be integrated.}

\item{angl}{An \code{\link{anglemania_object-class}} previously generated
using \code{\link{create_anglemania_object}} and \code{\link{anglemania}}.
It is important that the \code{dataset_key} and \code{batch_key} are
correctly set in the \code{anglemania_object}.}

\item{int_order}{An optional data frame specifying the integration order
of samples within the Seurat list. See the \code{sample.tree} argument
in \code{\link[Seurat]{IntegrateData}} for more details. If not
provided, Seurat will construct the integration order using
hierarchical clustering. Default is \code{NULL}.}

\item{process}{Logical value indicating whether to further process the
data after integration (i.e., scale it, run PCA, and compute UMAP
embeddings). Default is \code{TRUE}.}

\item{verbose}{Logical value indicating whether to display progress
messages during integration. Default is \code{FALSE}.}
}
\value{
A \code{\link[Seurat]{Seurat}} object containing the integrated
data. The default assay is set to \code{"integrated"}.
}
\description{
\code{integrate_by_features} integrates samples or batches within a Seurat
object using canonical correlation analysis (CCA) based on a set of
selected features (genes). The function utilizes an \code{anglemania_object} to
extract anglemania genes and handles the integration process, including
optional downstream processing steps such as scaling, PCA, and UMAP
visualization.
}
\details{
The function performs the following steps:
\enumerate{
\item \strong{Batch Key Addition}: Adds a unique batch key to the Seurat
object's metadata to distinguish different batches or samples. Batch
key is set to the \code{anglemania_object}'s \code{batch_key}.
\item \strong{Splitting}: Splits the Seurat object into a list of Seurat
objects based on the batch key.
\item \strong{Integration}: Calls \code{\link{integrate_seurat_list}} to
integrate the list of Seurat objects using the features extracted
from the \code{anglemania_object}.
}

The integration is performed using Seurat's CCA-based methods, and
parameters are adjusted based on the smallest dataset to ensure
compatibility with small sample sizes (e.g., metacells or SEACells). If
\code{process = TRUE}, the function will also scale the data, run PCA,
and compute UMAP embeddings.
}
\examples{
# Integrate samples using anglemania_object 
# Automatically reads the batch key from anglemania_object
#  splits the seurat object into batches and integrates them
#  using CCA integration and anglemania genes previously extracted
#  with anglemania() or select_genes()
sce <- sce_example()
se <- Seurat::as.Seurat(sce, data = "counts")
se <- SeuratObject::RenameAssays(se, "originalexp", "RNA")
angl <- create_anglemania_object(se, batch_key = "batch")
angl <- anglemania(
  angl, 
  zscore_mean_threshold = 2,
  zscore_sn_threshold = 2
)
options(future.globals.maxSize = 4000 * 1024^2)
integrated_object <- integrate_by_features(se, angl)
}
\seealso{
\code{\link{create_anglemania_object}},
\code{\link{anglemania}},
\code{\link{get_anglemania_genes}},
\code{\link{integrate_seurat_list}},
\code{\link[Seurat]{IntegrateData}},
\code{\link[Seurat]{FindIntegrationAnchors}}
}
