% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anglemanise_utils.R
\name{sparse_to_fbm}
\alias{sparse_to_fbm}
\title{Convert a sparse matrix into a file-backed matrix
(\code{\link[bigstatsr]{FBM}})}
\usage{
sparse_to_fbm(s_mat)
}
\arguments{
\item{s_mat}{A sparse matrix.}
}
\value{
An \code{\link[bigstatsr]{FBM}} object from the \pkg{bigstatsr}
package.
}
\description{
Converts a sparse matrix into an \code{\link[bigstatsr]{FBM}} with efficient
memory usage.
}
\examples{
s_mat <- Matrix::rsparsematrix(nrow = 10, ncol = 5, density = 0.3)
# Convert the sparse matrix to an FBM using your function
fbm_mat <- sparse_to_fbm(s_mat)
fbm_mat
}
